backend/requirements.txt
authlib
httpx
psycopg2-binary
fastapi
itsdangerous
uvicorn
python-dotenv
yahooquery
yfinance
fredapi
statsmodels
